+++ R/f7-update-inputs.R
@@ -202,7 +202,7 @@ updateF7Checkbox <- function(inputId, label = NULL, value = NULL,
 #'  ui <- f7Page(
 #'    f7SingleLayout(
 #'     navbar = f7Navbar(title = "updateF7Text"),
-#'     f7Fab("trigger", "Click me"),
+#'     f7Block(f7Button("trigger", "Click me")),
 #'     f7Text(
 #'      inputId = "text",
 #'      label = "Caption",