+++ inst/framework7-5.3.0/my-app.js
@@ -487,4 +487,19 @@ $(function () {
     });
   }
 
+  // validate inputs (see f7ValidateInput)
+  Shiny.addCustomMessageHandler('validate', function(message) {
+    $('#' + message.target)
+      .attr('required','')
+      .attr('validate', '')
+      .attr('pattern', message.pattern)
+      .attr('data-error-message', message.error);
+
+    $('#' + message.target).closest('.item-content.item-input').addClass('item-input-with-info');
+    var infoTag;
+    if (message.info !== undefined) {
+      infoTag = '<div class = "item-input-info">' + message.info + '</div>';
+    }
+    $('#' + message.target).parent().append(infoTag);
+  });
 });